add validation on raising when theres no bet etc

+++ card_utils/games/poker/community/omaha/plo/game_state.py
@@ -13,4 +13,4 @@ class PLOGameState(OmahaGameState):
 
         :return: (int)
         """
-        return 2 * self.amount_to_call + self.pot.total_money
+        return int(2 * self.amount_to_call + self.pot.total_money)
